+++ CHANGELOG.md
@@ -14,6 +14,7 @@ of these changes a reindex of Elasticsearch is required. This can be started by 
 - API add tags endpoint now returns the added tags.
   [CATS-1053](https://opensource.ncsa.illinois.edu/jira/browse/CATS-1053)
 - Ability to search by creator name and email address for all resources.
+- Ability to filter the files and folders in a dataset when sorting is enabled.
   
 ### Fixed
 - Ability to delete tags on file page.